chore: move to esm
