+++ src/annotations/index.ts
@@ -1,24 +1,24 @@
-import accessAnnotation from "./access";
-import aliasAnnotation from "./alias";
-import authorAnnotation from "./author";
-import contentAnnotation from "./content";
-import deprecatedAnnotation from "./deprecated";
-import exampleAnnotation from "./example";
-import groupAnnotation from "./group";
-import groupDescriptionAnnotation from "./groupDescription";
-import ignoreAnnotation from "./ignore";
-import linkAnnotation from "./link";
-import nameAnnotation from "./name";
-import outputAnnotation from "./output";
-import parameterAnnotation from "./parameter";
-import propertyAnnotation from "./property";
-import requireAnnotation from "./require";
-import returnAnnotation from "./return";
-import seeAnnotation from "./see";
-import sinceAnnotation from "./since";
-import throwAnnotation from "./throw";
-import todoAnnotation from "./todo";
-import typeAnnotation from "./type";
+import accessAnnotation from "./access.js";
+import aliasAnnotation from "./alias.js";
+import authorAnnotation from "./author.js";
+import contentAnnotation from "./content.js";
+import deprecatedAnnotation from "./deprecated.js";
+import exampleAnnotation from "./example.js";
+import groupAnnotation from "./group.js";
+import groupDescriptionAnnotation from "./groupDescription.js";
+import ignoreAnnotation from "./ignore.js";
+import linkAnnotation from "./link.js";
+import nameAnnotation from "./name.js";
+import outputAnnotation from "./output.js";
+import parameterAnnotation from "./parameter.js";
+import propertyAnnotation from "./property.js";
+import requireAnnotation from "./require.js";
+import returnAnnotation from "./return.js";
+import seeAnnotation from "./see.js";
+import sinceAnnotation from "./since.js";
+import throwAnnotation from "./throw.js";
+import todoAnnotation from "./todo.js";
+import typeAnnotation from "./type.js";
 
 export default [
   accessAnnotation,
