+++ src/index.ts
@@ -1,2 +1,17 @@
-export * from "./types";
-export * from "./sassdoc-parser";
+export * from "./sassdoc-parser.js";
+export type {
+  Context,
+  ContextType,
+  Example,
+  File,
+  Link,
+  Parameter,
+  ParseResult,
+  Property,
+  Range,
+  Require,
+  Return,
+  See,
+  Since,
+  UsedBy,
+} from "./types.js";
