+++ src/sorter.ts
@@ -1,4 +1,4 @@
-import { ParseResult } from "./types";
+import type { ParseResult } from "./types.js";
 
 export default function sort(data: Array<ParseResult>) {
   return data.sort((a, b) => {
